requirements.txt
google-auth-oauthlib
google-api-python-client
requests
google-cloud-language